Refactor U2F trait into usb::U2f.
